feat!: refine Monero crate and add Monero-LWS, Monero balance, and start of Monero transfer functionality
Add a hash module under Monero to handle keccak256 hash.

Add new structs to handle Monero keys in a more comprehensive way.
Functions written for structs related to handling Monero keys. The PrivateKey and
PublicKey structs hold a valid private or public key respectively. The MoneroPrivateKeys
and MoneroPublicKeys structs contains the information about the both the relevant spend
and view keys. Add tests for the Monero key structs.

Add new structs to handle Monero addresses in a more comprehensive way.
Rename MoneroFormat to AddressType. Add subaddress derivation and address generation.

Facilitate getting info from Monero-LWS type endpoints.

Add key image and key derivation functionality for Monero. Add tests for Monero key image.

Add the monero-generators crate as a dependency to provide the
hash_to_point function.

Able to get the balance of a Monero wallet.

Add dependency on zeroize and use of it for first time in monero crate

Add some functionality to complete a Monero transfer, but in the current state unable to create a valid transaction.
Able to discern unspent outs belonging to wallet. Add structs needed to handle a Transaction.
Add structs to handle rct_types. Handle ring signatures. Handle CLSAG. Functionality for bulletproof plus. 
Add functions needed to initiate, create, and sign a monero transaction. Derive view tags.
Be able to serialize the Monero data to json and data using the same protocol as the Monero source code, added DoSerialize trait for Monero and a SerializedArchive struct. Handle varint properly, added tests for varint.
Attempted to replicate the code in Monero codebase but currently there is something off.
An example of a transfer using stagenet was added, currently this example will not be able to successfully post a transaction to the blockchain but goes through all the steps and itself can determine it did not create a valid transaction so will not even attempt to submit it to the blockchain. 

BREAKING_CHANGE: Rename new_from_hd_keys to simply from_hd_key as this is more idiomatic. Monero keys and address handling changed significantly. 
Closes: #47, #29, #22